Add table support in report class
